Add DiscoverySystem and MutableDiscoverySystem interfaces
